Refeito o cálculo e ajuste na tabela do Alimento, na tela de Refeição

+++ Projeto_CDS/src/br/edu/ifsc/cds/classes/domain/Strategy/ColherCha.java
@@ -11,7 +11,7 @@ public class ColherCha implements UnidadeMedida {
 	}
 
 	@Override
-	public float Totalcalorias(int qtd, float calorias) {
+	public float Totalcalorias(float qtd, float calorias) {
 		return (qtd * 5) + calorias;
 	}
