+++ Projeto_CDS/src/br/edu/ifsc/cds/classes/domain/Strategy/ColherSopa.java
@@ -11,7 +11,7 @@ public class ColherSopa implements UnidadeMedida {
 	}
 
 	@Override
-	public float Totalcalorias(int qtd, float calorias) {
+	public float Totalcalorias(float qtd, float calorias) {
 		return (qtd * 10) + calorias;
 	}
